Partial manual review of all the diffs (fixed imports). In addition, I extracted default_eval and helpers to an eval defaults module.
