+++ experiments/dclm/exp433_dclm_run.py
@@ -4,10 +4,10 @@
 from fray.cluster import ResourceConfig
 from levanter.data.text import TextLmDatasetFormat
 from marin.defaults import SimpleTrainConfig, default_tokenize, default_train
+from marin.defaults.evals import default_eval
 from marin.execution.executor import executor_main
 from marin.processing.tokenize.data_configs import lm_mixture_data_config
 
-from experiments.evals.evals import default_eval
 from experiments.evals.task_configs import CORE_TASKS_PLUS_MMLU
 from experiments.llama import LlamaConfig
 from experiments.pretraining_datasets.dclm import DCLM_BASELINE_ONLY_MIXTURE, DCLM_MIXTURE_WEIGHTS
