+++ experiments/grug/moe/launch.py
@@ -21,7 +21,7 @@ from levanter.optim import OptimizerConfig
 from levanter.tracker import TrackerConfig
 from levanter.tracker.wandb import WandbConfig
 from levanter.trainer import TrainerConfig
-from marin.defaults import default_validation_sets
+from marin.defaults import default_tokenize, default_validation_sets
 from marin.execution.executor import executor_main
 from marin.execution.types import ExecutorStep, this_output_path, versioned
 from marin.processing.tokenize import add_validation_sets_to_mixture
@@ -33,7 +33,6 @@ from experiments.grug.moe.model import GrugModelConfig
 from experiments.grug.moe.train import GrugEvalConfig, GrugRunConfig, GrugTrainerConfig, run_grug
 from experiments.llama import llama3_tokenizer
 from experiments.pretraining_datasets import nemotron_mix
-from experiments.tokenization import default_tokenize
 
 
 @dataclass(frozen=True)
