+++ experiments/midtraining_datasets.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from marin.datakit.download.huggingface import DownloadConfig, download_hf
-from marin.defaults import default_download
+from marin.defaults import default_download, default_tokenize
 from marin.execution import versioned
 from marin.execution.types import ExecutorStep, this_output_path
 from marin.processing.tokenize import lm_mixture_data_config
@@ -14,7 +14,6 @@ from marin.transform.medical.lavita_to_dolma import LavitaToDolmaConfig, convert
 
 from experiments.common_pile.tokenize_common_pile import stackv2_edu_filtered
 from experiments.llama import llama3_tokenizer
-from experiments.tokenization import default_tokenize
 
 finemath_commit_hash = "8f233cf"
 finemath = ExecutorStep(
