+++ experiments/references/reference_training_pipeline.py
@@ -17,7 +17,7 @@ from fray.cluster import ResourceConfig
 from levanter.data.text import ChatLmDatasetFormat
 from levanter.optim import AdamConfig
 from levanter.tracker.wandb import WandbConfig
-from marin.defaults import default_validation_sets
+from marin.defaults import default_tokenize, default_validation_sets
 from marin.execution.types import this_output_path
 from marin.processing.tokenize import add_validation_sets_to_mixture
 from marin.processing.tokenize.data_configs import lm_varying_mixture_data_config
@@ -29,7 +29,6 @@ from experiments.marin_models import marin_tokenizer
 from experiments.posttrain.instruction_datasets import get_instruction_dataset
 from experiments.pretraining_datasets.dclm import dclm_components_llama3
 from experiments.pretraining_datasets.dolmino import tokenize_dolmino
-from experiments.tokenization import default_tokenize
 
 # --- Model: 600M Grug ---
 model = GrugModelConfig(
