--- experiments/simple_train_config.py
@@ -1,4 +1,0 @@
-# Copyright The Marin Authors
-# SPDX-License-Identifier: Apache-2.0
-
-from marin.defaults import SimpleTrainConfig  # noqa: F401
