+++ experiments/tutorials/train_tiny_sweep_dclm_tpu.py
@@ -8,14 +8,12 @@ on a 30M parameter DCLM model using TPU hardware.
 import dataclasses
 
 from fray.cluster import ResourceConfig
-from marin.defaults import CORE_TASKS
+from marin.defaults import CORE_TASKS, SimpleTrainConfig, default_train
 from marin.execution.executor import executor_main
 from marin.execution.types import versioned
 
-from experiments.defaults import default_train
 from experiments.llama import llama_30m
 from experiments.pretraining_datasets.dclm import dclm_mixture_config_llama3
-from experiments.simple_train_config import SimpleTrainConfig
 
 RESOURCES = ResourceConfig.with_tpu("v4-8")
 EVALS = CORE_TASKS
