+++ lib/marin/src/marin/defaults/__init__.py
@@ -2,5 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .config import SimpleDPOConfig, SimpleSFTConfig, SimpleTrainConfig
+from .evals import (
+    ACTIVE_DATASETS,
+    ALL_UNCHEATABLE_EVAL_DATASETS,
+    EVAL_DEPENDENCY_GROUPS,
+    default_eval,
+    uncheatable_eval_tokenized,
+)
 from .step import CORE_TASKS, default_dpo, default_sft, default_train, default_validation_sets
 from .tokenize import default_download, default_tokenize
